Move gravity-aware IMU factors into dedicated headers
Per review, ImuFactorWithGravityT now lives in ImuFactorWithGravity.h/.cpp
and CombinedImuFactorWithGravityT in CombinedImuFactorWithGravity.h/.cpp,
with their tests split into matching testImuFactorWithGravity.cpp and
testCombinedImuFactorWithGravity.cpp. The shared 15-dof error assembly
helper moves from CombinedImuFactor.cpp into internal::combinedImuError in
CombinedImuFactor.h since both translation units use it. No functional
change.
