+++ gtsam/navigation/ImuFactor.cpp
@@ -210,74 +210,5 @@ template GTSAM_EXPORT std::ostream& operator<<<PreintegratedImuMeasurementsT<Man
 template GTSAM_EXPORT std::ostream& operator<<<PreintegratedImuMeasurementsT<TangentPreintegration>>(
     std::ostream& os, const ImuFactor2T<PreintegratedImuMeasurementsT<TangentPreintegration>>& f);
 
-//------------------------------------------------------------------------------
-// ImuFactorWithGravityT methods
-//------------------------------------------------------------------------------
-template <class PIM, class GRAVITY>
-std::ostream& operator<<(std::ostream& os, const ImuFactorWithGravityT<PIM, GRAVITY>& f) {
-  f.preintegratedMeasurements().print("preintegrated measurements:\n");
-  if (internal::GravityParametrization<GRAVITY>::usesMagnitude)
-    os << "  gravity magnitude: " << f.gravityMagnitude() << "\n";
-  os << "  noise model sigmas: " << f.noiseModel()->sigmas().transpose();
-  return os;
-}
-
-//------------------------------------------------------------------------------
-template <class PIM, class GRAVITY>
-void ImuFactorWithGravityT<PIM, GRAVITY>::print(const string& s,
-    const KeyFormatter& keyFormatter) const {
-  cout << (s.empty() ? s : s + "\n") << "ImuFactorWithGravity("
-       << keyFormatter(this->template key<1>()) << "," << keyFormatter(this->template key<2>())
-       << "," << keyFormatter(this->template key<3>()) << "," << keyFormatter(this->template key<4>())
-       << "," << keyFormatter(this->template key<5>()) << "," << keyFormatter(this->template key<6>())
-       << ")\n";
-  cout << *this << endl;
-}
-
-//------------------------------------------------------------------------------
-template <class PIM, class GRAVITY>
-bool ImuFactorWithGravityT<PIM, GRAVITY>::equals(const NonlinearFactor& other,
-    double tol) const {
-  const This *e = dynamic_cast<const This*>(&other);
-  return e != nullptr && Base::equals(*e, tol) && pim_.equals(e->pim_, tol) &&
-         (!internal::GravityParametrization<GRAVITY>::usesMagnitude ||
-          std::abs(gravityMagnitude_ - e->gravityMagnitude_) < tol);
-}
-
-//------------------------------------------------------------------------------
-template <class PIM, class GRAVITY>
-Vector ImuFactorWithGravityT<PIM, GRAVITY>::evaluateError(const Pose3& pose_i,
-    const Vector3& vel_i, const Pose3& pose_j, const Vector3& vel_j,
-    const imuBias::ConstantBias& bias_i, const GRAVITY& gravity,
-    OptionalMatrixType H1, OptionalMatrixType H2, OptionalMatrixType H3,
-    OptionalMatrixType H4, OptionalMatrixType H5, OptionalMatrixType H6) const {
-  typedef internal::GravityParametrization<GRAVITY> Parametrization;
-  Eigen::Matrix<double, 3, Parametrization::dimension> D_gvec_gravity;
-  const Vector3 n_gravity = Parametrization::vector(
-      gravity, gravityMagnitude_, H6 ? &D_gvec_gravity : nullptr);
-  Matrix93 D_r_gvec;
-  const Vector r = pim_.computeErrorAndJacobians(
-      pose_i, vel_i, pose_j, vel_j, bias_i, n_gravity, H1, H2, H3, H4, H5,
-      H6 ? &D_r_gvec : nullptr);
-  if (H6) *H6 = D_r_gvec * D_gvec_gravity;
-  return r;
-}
-
-//------------------------------------------------------------------------------
-// ImuFactorWithGravityT instantiations
-template class GTSAM_EXPORT ImuFactorWithGravityT<PreintegratedImuMeasurementsT<ManifoldPreintegration>, Unit3>;
-template class GTSAM_EXPORT ImuFactorWithGravityT<PreintegratedImuMeasurementsT<TangentPreintegration>, Unit3>;
-template class GTSAM_EXPORT ImuFactorWithGravityT<PreintegratedImuMeasurementsT<ManifoldPreintegration>, Point3>;
-template class GTSAM_EXPORT ImuFactorWithGravityT<PreintegratedImuMeasurementsT<TangentPreintegration>, Point3>;
-
-template GTSAM_EXPORT std::ostream& operator<<(
-    std::ostream& os, const ImuFactorWithGravityT<PreintegratedImuMeasurementsT<ManifoldPreintegration>, Unit3>& f);
-template GTSAM_EXPORT std::ostream& operator<<(
-    std::ostream& os, const ImuFactorWithGravityT<PreintegratedImuMeasurementsT<TangentPreintegration>, Unit3>& f);
-template GTSAM_EXPORT std::ostream& operator<<(
-    std::ostream& os, const ImuFactorWithGravityT<PreintegratedImuMeasurementsT<ManifoldPreintegration>, Point3>& f);
-template GTSAM_EXPORT std::ostream& operator<<(
-    std::ostream& os, const ImuFactorWithGravityT<PreintegratedImuMeasurementsT<TangentPreintegration>, Point3>& f);
-
 }
 // namespace gtsam
